pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["boison <boison88@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
django-bootstrap5 = "^23.3"
whitenoise = "^6.5.0"
dj-database-url = "^2.0.0"
django-crispy-forms = "^2.0"
crispy-bootstrap5 = "^0.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
